[TASK] Move logic from AbstractRepository into CacheTableRepository
