+++ Classes/Domain/Repository/CacheTableRepository.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Aoe\Cachemgm\Domain\Repository;
 
+use TYPO3\CMS\Core\Database\ConnectionPool;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+
 /***************************************************************
  *  Copyright notice
  *
@@ -28,6 +31,15 @@ namespace Aoe\Cachemgm\Domain\Repository;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
-class CacheTableRepository extends AbstractRepository
+class CacheTableRepository
 {
+    public function countRowsInTable(string $table): int
+    {
+        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable($table);
+        return $queryBuilder
+            ->count('*')
+            ->from($table)
+            ->executeQuery()
+            ->fetchOne();
+    }
 }
